+++ index-handlers/src/main/java/no/sikt/nva/nvi/index/model/Contexts.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.nvi.index;
+package no.sikt.nva.nvi.index.model;
 
 public final class Contexts {
