+++ index-handlers/src/test/resources/resourceSample.json
